additional functions in kernel.js
